Separate OpenAPI schema inspection from Onshape annotations
